refactor: usar os hooks customizados pra chamadas na api
